Implement flip().
This passes a callback from the Game View to the Lua bridge, so that flip can
call back in to SDL.

While there's some overlap with ::render(), the latter has some better
knowledge (like whether the game is paused) which makes it hard to share the
code directly with ::flip().

While here, rearrange the init code a little so that things are all started
before the game's Lua code is first loaded; Picopolis calls flip() immediately
rather than waiting for _init or _update.